// panels/projects/coach/src/app/models/excercise.model.ts
// Generated by https://quicktype.io

export interface Exercise {
  _id: string;
  file: string;
  mimetype: string;
  name: string;
  description: string;
  steps: Step[];
  exCategory: string;
  exSubCategory: string;
  coach: string;
  __v: number;
}

export interface Step {
  title: string;
  content: string;
  _id: string;
}